Extract functions that call into user space into CommitEffects
